third_party/blink/web_tests/external/wpt/css/css-text/parsing/white-space-shorthand-text-wrap-expected.txt

This is a testharness.js-based test.
[FAIL] `white-space` should not overwrite previous `text-wrap-style: balance`
  assert_equals: expected "balance" but got "wrap"
[FAIL] `white-space` should overwrite previous `text-wrap-mode: nowrap`
  assert_equals: expected "balance" but got "wrap"
[FAIL] `white-space` should overwrite `text-wrap-mode` on the parent
  assert_equals: expected "balance" but got "wrap"
Harness: the test ran to completion.
